Factor out some shared logic
