+++ tests/Aspire.Azure.AI.OpenAI.Tests/AspireAzureOpenAIClientBuilderEmbeddingGeneratorExtensionsTests.cs
@@ -157,7 +157,7 @@ public class AspireAzureOpenAIClientBuilderEmbeddingGeneratorExtensionsTests
                 host.Services.GetRequiredService<IEmbeddingGenerator<string, Embedding<float>>>();
         });
 
-        Assert.StartsWith($"An {nameof(IEmbeddingGenerator<string, Embedding<float>>)} could not be configured", ex.Message);
+        Assert.StartsWith("The deployment could not be determined", ex.Message);
     }
 
     [Theory]
