+++ tests/Aspire.OpenAI.Tests/AspireOpenAIClientBuilderChatClientExtensionsTests.cs
@@ -158,7 +158,7 @@ public class AspireOpenAIClientBuilderChatClientExtensionsTests
                 host.Services.GetRequiredService<IChatClient>();
         });
 
-        Assert.StartsWith($"An {nameof(IChatClient)} could not be configured", ex.Message);
+        Assert.StartsWith("The deployment could not be determined", ex.Message);
     }
 
     [Theory]
